+++ Cargo.toml
@@ -16,3 +16,4 @@ iced_core = "0.2"
 iced_web = "0.2"
 iced_wgpu = "0.2"
 iced_winit = "0.1"
+iced_futures = "0.1.0"
